(scratchbox2) Add options to not set any default dirs for ld.so
Adds a new option "--nodefaultdirs" to ld.so.

This option disables searching from the default directories
(/lib,/usr/lib) and also disables use of ld.so's search
cache.

This is recommended for correct operation with Scratchbox 2;
without this, libraries may be incorrectly loaded from the
host environment (such libraries that are missing from
the "scratchboxed" environment but are present in the host).